Revert "move E.myguid"
This reverts commit 66ab64abc6f8b985838bd46600f6c9a851c9eab8.

We need E.myguid earlier.
Initialization has a call to `E:InitializeModules()`.
This in turn calls `CH:Initialize()`.
Which calls `CH:DisplayChatHistory()`.

And `CH:DisplayChatHistory` has a delay until information about
ourselves is known:

```lua
  if not CH:GetPlayerInfoByGUID(E.myguid) then
    E:Delay(0.1, CH.DisplayChatHistory)
    return
  end
```

and that would loop infinitely -- unless the chat module was disabled.

If ran on a fresh character, or with its character-specific ElvUI SVars
deleted (which are only chat cvars), it wouldn't loop, because of this
guard right above the above code:

```lua
  local data = ElvCharacterDB.ChatHistoryLog
  if not (data and next(data)) then return end
```

which short-circuited the whole function.